ShaeIt ТЗ 16 gateway server test-2 корректировка 1-2
