+++ gateway/src/main/java/ru.practicum.shareit/Constants.java
@@ -1,0 +1,5 @@
+package ru.practicum.shareit;
+
+public class Constants {
+    public static final String USER_ID = "X-Sharer-User-Id";
+}
