+++ gateway/src/main/java/ru.practicum.shareit/State.java
@@ -9,6 +9,3 @@ public enum State {
     REJECTED;
 }
 
-    public class Constants {
-        public static final String USER_ID = "X-Sharer-User-Id";
-    }
