--- gateway/src/main/java/ru.practicum.shareit/expections/UnsupportedState.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.expections;
-
-public class UnsupportedState extends RuntimeException {
-    public UnsupportedState(final String message) {
-        super(message);
-    }
-}
