+++ server/src/main/java/ru.practicum.shareit/expections/ErrorHandler.java
@@ -1,8 +1,8 @@
 package ru.practicum.shareit.expections;
 
 import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -16,13 +16,6 @@ public class ErrorHandler {
     @ResponseStatus(HttpStatus.NOT_FOUND) //404
     public ErrorResponse handleNotFoundException(final NotFoundException exception) {
         log.debug("handleNotFoundException Получен статус 404 Not found {}", exception.getMessage(), exception);
-        return new ErrorResponse(exception.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST) // 400
-    public ErrorResponse handleValidationException(final ValidationException exception) {
-        log.debug("handleValidationException Получен статус 400 Bad request {}", exception.getMessage(), exception);
         return new ErrorResponse(exception.getMessage());
     }
 
@@ -47,14 +40,6 @@ public class ErrorHandler {
         return new ErrorResponse("http:400 Искомый объект не найден при первичной проверке.", exception.getMessage());
     }
 
-    @ResponseBody
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST) // 400
-    public ErrorResponse handleMethodArgumentNotValidException(final MethodArgumentNotValidException exception) {
-        log.debug("handleMethodArgumentNotValidException Искомый объект не найден. Возврат код 400 {}", exception.getMessage());
-        return new ErrorResponse("http:400 Искомый объект не найден при первичной проверке.", exception.getMessage());
-    }
-
     @ExceptionHandler
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR) // 500
     public ErrorResponse handleServerError(final Exception exception) {
@@ -74,10 +59,4 @@ public class ErrorHandler {
         return new ErrorResponse(exception.getMessage());
     }
 
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR) // 500
-    public ErrorResponse handleMethodArgumentNotValidException(
-            final ru.practicum.shareit.expections.MethodArgumentNotValidException exception) {
-        return new ErrorResponse(exception.getMessage());
-    }
 }
