--- server/src/test/java/ru.practicum.shareit/expections/MethodArgumentNotValidExceptionTest.java
@@ -1,23 +1,0 @@
-package ru.practicum.shareit.expections;
-
-import lombok.RequiredArgsConstructor;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional
-@SpringBootTest
-@RequiredArgsConstructor(onConstructor_ = @Autowired)
-class MethodArgumentNotValidExceptionTest {
-
-    private final ErrorHandler errorHandler;
-
-    @Test
-    void methodArgumentNotValidExceptionTests() {
-        ErrorResponse error = errorHandler.handleMethodArgumentNotValidException(new MethodArgumentNotValidException("message"));
-        Assertions.assertEquals(error.getError(), "500 INTERNAL_SERVER_ERROR");
-    }
-
-}
