+++ server/src/test/java/ru.practicum.shareit/item/ItemControllerTest.java
@@ -6,7 +6,7 @@ import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static ru.practicum.shareit.service.MyConstants.USER_ID;
+import static ru.practicum.shareit.Constants.USER_ID;
 
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
