+++ server/src/test/java/ru.practicum.shareit/request/ItemRequestControllerTest.java
@@ -10,7 +10,7 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static ru.practicum.shareit.service.MyConstants.USER_ID;
+import static ru.practicum.shareit.Constants.USER_ID;
 
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
